refactor: 개별 예외에서 UnsupportedException 공용 예외로 변경 #11
