+++ src/main/java/com/dh/ondot/member/core/OauthProviderConverter.java
@@ -1,10 +1,11 @@
 package com.dh.ondot.member.core;
 
-import com.dh.ondot.member.core.exception.UnsupportedSocialLoginException;
+import com.dh.ondot.core.exception.UnsupportedException;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 import com.dh.ondot.member.domain.OauthProvider;
 
+import static com.dh.ondot.core.exception.ErrorCode.UNSUPPORTED_SOCIAL_LOGIN;
 import static java.util.Locale.ENGLISH;
 
 @Component
@@ -14,7 +15,7 @@ public class OauthProviderConverter implements Converter<String, OauthProvider> 
         try {
             return OauthProvider.valueOf(type.toUpperCase(ENGLISH));
         } catch (IllegalArgumentException e) {
-            throw new UnsupportedSocialLoginException(type);
+            throw new UnsupportedException(UNSUPPORTED_SOCIAL_LOGIN, type);
         }
     }
 }
