+++ src/main/java/com/dh/ondot/member/infra/oauth/OauthApiFactoryImpl.java
@@ -1,9 +1,9 @@
 package com.dh.ondot.member.infra.oauth;
 
-import com.dh.ondot.member.core.exception.UnsupportedSocialLoginException;
 import com.dh.ondot.member.domain.OauthApi;
 import com.dh.ondot.member.domain.OauthApiFactory;
 import com.dh.ondot.member.domain.OauthProvider;
+import com.dh.ondot.core.exception.UnsupportedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -11,6 +11,8 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+
+import static com.dh.ondot.core.exception.ErrorCode.UNSUPPORTED_SOCIAL_LOGIN;
 
 @Component
 @RequiredArgsConstructor
@@ -28,6 +30,6 @@ public class OauthApiFactoryImpl implements OauthApiFactory {
     @Override
     public OauthApi getOauthApi(OauthProvider provider) {
         return Optional.ofNullable(oauthApiMap.get(provider))
-                .orElseThrow(() -> new UnsupportedSocialLoginException(provider.name()));
+                .orElseThrow(() -> new UnsupportedException(UNSUPPORTED_SOCIAL_LOGIN, provider.name()));
     }
 }
