+++ src/main/java/com/dh/ondot/schedule/domain/enums/RingTone.java
@@ -1,7 +1,7 @@
 package com.dh.ondot.schedule.domain.enums;
 
 import com.dh.ondot.core.exception.ErrorCode;
-import com.dh.ondot.schedule.core.exception.UnsupportedException;
+import com.dh.ondot.core.exception.UnsupportedException;
 
 import java.util.Locale;
